chore(desktop): move inline bilingual copy into catalogs
Typed locale catalogs are the mandated home for user-visible copy;
inline locale ternaries scattered through components bypass that and
make copy audits miss strings. This is the mechanical consolidation
wave: zh and en strings are moved verbatim into UiCatalog-shaped
tables with no behavior or string changes. The one exception is the
peer route state catalog, which loses a sixth key that the closed
five-value union could never render.

Generated-by: Claude Code
