+++ apps/desktop/src/main/notifications-policy.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import type { UiLocale } from '@maka/core/ui-locale';
+import type { UiCatalog, UiLocale } from '@maka/core/ui-locale';
 
 /**
  * Pure decision + copy helpers for desktop run-completion notifications.
@@ -77,24 +77,26 @@ export interface RunNotificationCopy {
  * could not supply a session name / reply preview (e.g. an untitled
  * session or a tool-only turn with no assistant text).
  */
+const RUN_NOTIFICATION_COPY = {
+  'zh-CN': {
+    errored: { title: '任务出错', body: '本轮回答未能完成，点击查看详情。' },
+    completed: { title: '回答已生成', body: 'Maka 已完成本轮回答，点击查看。' },
+  },
+  'zh-TW': {
+    errored: { title: '任務發生錯誤', body: '本次回答未能完成，按一下以檢視詳細資料。' },
+    completed: { title: '回答已產生', body: 'Maka 已完成本次回答，按一下以檢視。' },
+  },
+  en: {
+    errored: { title: 'Conversation error', body: 'This response did not finish. Click to view details.' },
+    completed: { title: 'Response ready', body: 'Maka finished this response. Click to view it.' },
+  },
+} satisfies UiCatalog<Record<RunNotificationKind, RunNotificationCopy>>;
+
 export function runNotificationCopy(
   kind: RunNotificationKind,
   locale: UiLocale,
 ): RunNotificationCopy {
-  if (locale === 'en') {
-    return kind === 'errored'
-      ? { title: 'Conversation error', body: 'This response did not finish. Click to view details.' }
-      : { title: 'Response ready', body: 'Maka finished this response. Click to view it.' };
-  }
-  if (locale === 'zh-TW') {
-    return kind === 'errored'
-      ? { title: '任務發生錯誤', body: '本次回答未能完成，按一下以檢視詳細資料。' }
-      : { title: '回答已產生', body: 'Maka 已完成本次回答，按一下以檢視。' };
-  }
-  if (kind === 'errored') {
-    return { title: '任务出错', body: '本轮回答未能完成，点击查看详情。' };
-  }
-  return { title: '回答已生成', body: 'Maka 已完成本轮回答，点击查看。' };
+  return RUN_NOTIFICATION_COPY[locale][kind];
 }
 
 /** Renderer-supplied content for a finished turn. Both fields are
