+++ apps/desktop/src/renderer/locales/settings-projects-copy.ts
@@ -121,6 +121,12 @@ export type SettingsProjectsCopy = {
     saveAndEnable: string;
     defaultBadge: string;
     experimentalBadge: string;
+    copyPeerId(value: string): string;
+    peerIdCopied: string;
+    peerIdCopyFailed: string;
+    peerPathDirect: string;
+    peerPathTransit: string;
+    peerPathTransportOther: string;
     defaultDisableHelp: string;
     unavailable: string;
     manage: string;
@@ -423,6 +429,12 @@ const SETTINGS_PROJECTS_COPY_BY_LOCALE = {
       saveAndEnable: '保存并启用',
       defaultBadge: '默认',
       experimentalBadge: '实验性',
+      copyPeerId: (value: string) => `复制完整 Peer ID：${value}`,
+      peerIdCopied: 'Peer ID 已复制',
+      peerIdCopyFailed: '无法复制 Peer ID',
+      peerPathDirect: '直连',
+      peerPathTransit: '成员转发',
+      peerPathTransportOther: '其他',
       defaultDisableHelp: '先选择另一个默认 Host，才能停用此 Host',
       unavailable: '无法连接',
       manage: '管理',
@@ -732,6 +744,12 @@ const SETTINGS_PROJECTS_COPY_BY_LOCALE = {
       credentialHelp: '在遠端機器使用 desktop-client preset 簽發',
       saveAndEnable: '儲存並啟用',
       defaultBadge: '預設',
+      copyPeerId: (value: string) => `複製完整 Peer ID：${value}`,
+      peerIdCopied: 'Peer ID 已複製',
+      peerIdCopyFailed: '無法複製 Peer ID',
+      peerPathDirect: '直接連線',
+      peerPathTransit: '成員轉送',
+      peerPathTransportOther: '其他',
       defaultDisableHelp: '先選擇另一個預設 Host，才能停用此 Host',
       unavailable: '無法連線',
       manage: '管理',
@@ -1043,6 +1061,12 @@ const SETTINGS_PROJECTS_COPY_BY_LOCALE = {
       saveAndEnable: 'Save and enable',
       defaultBadge: 'Default',
       experimentalBadge: 'Experimental',
+      copyPeerId: (value: string) => `Copy full Peer ID: ${value}`,
+      peerIdCopied: 'Peer ID copied',
+      peerIdCopyFailed: 'Could not copy Peer ID',
+      peerPathDirect: 'Direct',
+      peerPathTransit: 'Member transit',
+      peerPathTransportOther: 'Other',
       defaultDisableHelp: 'Choose another default Host before disabling this Host',
       unavailable: 'Unavailable',
       manage: 'Manage',
