+++ apps/desktop/src/renderer/mcp-catalog.ts
@@ -17,18 +17,15 @@
  * under the License.
  */
 
-import type { UiLocale } from '@maka/core/ui-locale';
+import type { UiCatalog, UiLocale } from '@maka/core/ui-locale';
 import type { McpServerConfig } from '@maka/core/mcp';
 
 export type McpCatalogId =
   | 'dingtalk' | 'feishu' | 'slack' | 'line' | 'notion' | 'macos-apps' | 'google-calendar'
   | 'figma' | 'vercel' | 'supabase' | 'filesystem' | 'memory' | 'playwright' | 'sequential-thinking';
 
-export type McpCatalogEntry = {
+type McpCatalogDefinition = {
   id: McpCatalogId;
-  name: string;
-  description: string;
-  category: string;
   /** Text fallback glyph for entries without a library brand mark (see
       mcp-brand-marks.tsx). Branded entries render their real mark instead, so
       they omit this. */
@@ -36,19 +33,14 @@ export type McpCatalogEntry = {
   aliases?: string[];
   config: McpServerConfig;
   setupRequired?: boolean;
-  setupLabel?: string;
   platform?: 'darwin';
 };
 
-export const MCP_CATALOG: McpCatalogEntry[] = [
+export const MCP_CATALOG: McpCatalogDefinition[] = [
   {
     id: 'dingtalk',
-    name: '钉钉',
-    description: '管理联系人、日历、待办与协作信息。',
-    category: '沟通协作',
     aliases: ['DingTalk'],
     setupRequired: true,
-    setupLabel: '需要 Client ID 与 Client Secret',
     config: {
       enabled: false,
       command: 'npx',
@@ -62,12 +54,8 @@ export const MCP_CATALOG: McpCatalogEntry[] = [
   },
   {
     id: 'feishu',
-    name: '飞书',
-    description: '访问飞书文档、日历、消息与 OpenAPI。',
-    category: '沟通协作',
     aliases: ['Feishu', 'Lark'],
     setupRequired: true,
-    setupLabel: '需要 App ID 与 App Secret',
     config: {
       enabled: false,
       command: 'npx',
@@ -77,12 +65,8 @@ export const MCP_CATALOG: McpCatalogEntry[] = [
   },
   {
     id: 'slack',
-    name: 'Slack',
-    description: '发送消息、管理频道并与 Slack workspace 协作。',
-    category: '沟通协作',
     mark: 'S',
     setupRequired: true,
-    setupLabel: '需要 Bot Token 与 Team ID',
     config: {
       enabled: false,
       command: 'npx',
@@ -92,12 +76,8 @@ export const MCP_CATALOG: McpCatalogEntry[] = [
   },
   {
     id: 'line',
-    name: 'LINE',
-    description: '通过 LINE Bot Messaging API 发送和管理消息。',
-    category: '沟通协作',
     mark: 'LINE',
     setupRequired: true,
-    setupLabel: '需要 Channel Access Token',
     config: {
       enabled: false,
       command: 'npx',
@@ -107,32 +87,21 @@ export const MCP_CATALOG: McpCatalogEntry[] = [
   },
   {
     id: 'notion',
-    name: 'Notion',
-    description: '搜索、读取和更新 Notion workspace。',
-    category: '知识与文档',
     mark: 'N',
     setupRequired: true,
-    setupLabel: '需要登录授权',
     config: { enabled: false, url: 'https://mcp.notion.com/mcp', transport: 'streamable-http', protocol: 'auto' },
   },
   {
     id: 'macos-apps',
-    name: 'macOS 应用',
-    description: '连接系统日历与提醒事项，并使用原生权限模型。',
-    category: '系统与效率',
     aliases: ['Apple', 'Calendar', 'Reminders'],
     platform: 'darwin',
     config: { enabled: true, command: 'npx', args: ['-y', 'mcp-server-apple-events@1.4.0'] },
   },
   {
     id: 'google-calendar',
-    name: 'Google 日历',
-    description: '管理日程、创建会议并查询空闲时间。',
-    category: '系统与效率',
     mark: '31',
     aliases: ['Google Calendar'],
     setupRequired: true,
-    setupLabel: '需要 OAuth credentials 文件',
     config: {
       enabled: false,
       command: 'npx',
@@ -142,12 +111,8 @@ export const MCP_CATALOG: McpCatalogEntry[] = [
   },
   {
     id: 'figma',
-    name: 'Figma',
-    description: '读取设计文件、组件与开发交付信息。',
-    category: '设计与开发',
     mark: 'F',
     setupRequired: true,
-    setupLabel: '需要 Personal Access Token',
     config: {
       enabled: false,
       command: 'npx',
@@ -157,32 +122,20 @@ export const MCP_CATALOG: McpCatalogEntry[] = [
   },
   {
     id: 'vercel',
-    name: 'Vercel',
-    description: '检查项目、部署状态、日志与平台文档。',
-    category: '设计与开发',
     mark: '▲',
     setupRequired: true,
-    setupLabel: '需要登录授权',
     config: { enabled: false, url: 'https://mcp.vercel.com', transport: 'streamable-http', protocol: 'auto' },
   },
   {
     id: 'supabase',
-    name: 'Supabase',
-    description: '管理数据库、项目配置、迁移与 Edge Functions。',
-    category: '设计与开发',
     mark: 'S',
     setupRequired: true,
-    setupLabel: '需要登录授权',
     config: { enabled: false, url: 'https://mcp.supabase.com/mcp', transport: 'streamable-http', protocol: 'auto' },
   },
   {
     id: 'filesystem',
-    name: '本地文件',
-    description: '在指定目录中安全地读取、写入和管理文件。',
-    category: '文件与知识',
     mark: 'FS',
     setupRequired: true,
-    setupLabel: '需要选择允许访问的目录',
     config: {
       enabled: false,
       command: 'npx',
@@ -191,25 +144,16 @@ export const MCP_CATALOG: McpCatalogEntry[] = [
   },
   {
     id: 'memory',
-    name: '持久记忆',
-    description: '用结构化知识图谱记住实体、关系和重要事实。',
-    category: '文件与知识',
     mark: 'M',
     config: { enabled: true, command: 'npx', args: ['-y', '@modelcontextprotocol/server-memory'] },
   },
   {
     id: 'playwright',
-    name: '浏览器自动化',
-    description: '让 Maka 通过 Playwright 读取和操作真实网页。',
-    category: '设计与开发',
     mark: 'PW',
     config: { enabled: true, command: 'npx', args: ['-y', '@playwright/mcp@latest'] },
   },
   {
     id: 'sequential-thinking',
-    name: '序列思考',
-    description: '为复杂问题提供可修正、可验证的结构化推理。',
-    category: '推理与规划',
     mark: 'ST',
     config: {
       enabled: true,
@@ -219,47 +163,68 @@ export const MCP_CATALOG: McpCatalogEntry[] = [
   },
 ];
 
-type McpCatalogLocalizedCopy = Pick<McpCatalogEntry, 'name' | 'description' | 'category'> & {
+type McpCatalogLocalizedCopy = {
+  name: string;
+  description: string;
+  category: string;
   setupLabel: string | undefined;
 };
 
-const MCP_CATALOG_ENGLISH_COPY: Record<McpCatalogId, McpCatalogLocalizedCopy> = {
-  dingtalk: { name: 'DingTalk', description: 'Manage contacts, calendars, tasks, and collaboration data.', category: 'Communication', setupLabel: 'Requires Client ID and Client Secret' },
-  feishu: { name: 'Feishu', description: 'Access Feishu documents, calendars, messages, and OpenAPI.', category: 'Communication', setupLabel: 'Requires App ID and App Secret' },
-  slack: { name: 'Slack', description: 'Send messages, manage channels, and collaborate in a Slack workspace.', category: 'Communication', setupLabel: 'Requires Bot Token and Team ID' },
-  line: { name: 'LINE', description: 'Send and manage messages through the LINE Bot Messaging API.', category: 'Communication', setupLabel: 'Requires Channel Access Token' },
-  notion: { name: 'Notion', description: 'Search, read, and update a Notion workspace.', category: 'Knowledge and documents', setupLabel: 'Requires sign-in authorization' },
-  'macos-apps': { name: 'macOS apps', description: 'Connect Calendar and Reminders through native system permissions.', category: 'System and productivity', setupLabel: undefined },
-  'google-calendar': { name: 'Google Calendar', description: 'Manage events, create meetings, and check availability.', category: 'System and productivity', setupLabel: 'Requires an OAuth credentials file' },
-  figma: { name: 'Figma', description: 'Read design files, components, and developer handoff data.', category: 'Design and development', setupLabel: 'Requires a Personal Access Token' },
-  vercel: { name: 'Vercel', description: 'Inspect projects, deployments, logs, and platform documentation.', category: 'Design and development', setupLabel: 'Requires sign-in authorization' },
-  supabase: { name: 'Supabase', description: 'Manage databases, project configuration, migrations, and Edge Functions.', category: 'Design and development', setupLabel: 'Requires sign-in authorization' },
-  filesystem: { name: 'Local files', description: 'Safely read, write, and manage files in selected directories.', category: 'Files and knowledge', setupLabel: 'Requires selecting allowed directories' },
-  memory: { name: 'Persistent memory', description: 'Remember entities, relationships, and important facts in a structured knowledge graph.', category: 'Files and knowledge', setupLabel: undefined },
-  playwright: { name: 'Browser automation', description: 'Let Maka read and operate real web pages through Playwright.', category: 'Design and development', setupLabel: undefined },
-  'sequential-thinking': { name: 'Sequential thinking', description: 'Provide revisable, verifiable structured reasoning for complex problems.', category: 'Reasoning and planning', setupLabel: undefined },
-};
+export type McpCatalogEntry = McpCatalogDefinition & McpCatalogLocalizedCopy;
 
-const MCP_CATALOG_TRADITIONAL_COPY: Record<McpCatalogId, McpCatalogLocalizedCopy> = {
-  dingtalk: { name: 'DingTalk', description: '管理聯絡人、行事曆、任務和協作資料。', category: '通訊', setupLabel: '需要 Client ID 和 Client Secret' },
-  feishu: { name: '飛書', description: '存取飛書文件、行事曆、訊息和 OpenAPI。', category: '通訊', setupLabel: '需要 App ID 和 App Secret' },
-  slack: { name: 'Slack', description: '傳送訊息、管理頻道，並在 Slack 工作區協作。', category: '通訊', setupLabel: '需要 Bot Token 和 Team ID' },
-  line: { name: 'LINE', description: '透過 LINE Bot Messaging API 傳送和管理訊息。', category: '通訊', setupLabel: '需要 Channel Access Token' },
-  notion: { name: 'Notion', description: '搜尋、讀取和更新 Notion 工作區。', category: '知識與文件', setupLabel: '需要登入授權' },
-  'macos-apps': { name: 'macOS 應用程式', description: '透過原生系統權限連接「行事曆」與「提醒事項」。', category: '系統與生產力', setupLabel: undefined },
-  'google-calendar': { name: 'Google 日曆', description: '管理活動、建立會議並檢查可用時間。', category: '系統與生產力', setupLabel: '需要 OAuth 憑證檔案' },
-  figma: { name: 'Figma', description: '讀取設計檔案、元件和開發交付資料。', category: '設計與開發', setupLabel: '需要 Personal Access Token' },
-  vercel: { name: 'Vercel', description: '檢查專案、部署、記錄和平台文件。', category: '設計與開發', setupLabel: '需要登入授權' },
-  supabase: { name: 'Supabase', description: '管理資料庫、專案設定、移轉和 Edge Functions。', category: '設計與開發', setupLabel: '需要登入授權' },
-  filesystem: { name: '本機檔案', description: '在選取的資料夾中安全地讀取、寫入和管理檔案。', category: '檔案與知識', setupLabel: '需要選擇允許存取的資料夾' },
-  memory: { name: '持久記憶', description: '使用結構化知識圖譜記住實體、關係和重要事實。', category: '檔案與知識', setupLabel: undefined },
-  playwright: { name: '瀏覽器自動化', description: '讓 Maka 透過 Playwright 讀取和操作實際網頁。', category: '設計與開發', setupLabel: undefined },
-  'sequential-thinking': { name: '循序思考', description: '為複雜問題提供可修正、可驗證的結構化推理。', category: '推理與規劃', setupLabel: undefined },
-};
+const MCP_CATALOG_COPY = {
+  'zh-CN': {
+    dingtalk: { name: '钉钉', description: '管理联系人、日历、待办与协作信息。', category: '沟通协作', setupLabel: '需要 Client ID 与 Client Secret' },
+    feishu: { name: '飞书', description: '访问飞书文档、日历、消息与 OpenAPI。', category: '沟通协作', setupLabel: '需要 App ID 与 App Secret' },
+    slack: { name: 'Slack', description: '发送消息、管理频道并与 Slack workspace 协作。', category: '沟通协作', setupLabel: '需要 Bot Token 与 Team ID' },
+    line: { name: 'LINE', description: '通过 LINE Bot Messaging API 发送和管理消息。', category: '沟通协作', setupLabel: '需要 Channel Access Token' },
+    notion: { name: 'Notion', description: '搜索、读取和更新 Notion workspace。', category: '知识与文档', setupLabel: '需要登录授权' },
+    'macos-apps': { name: 'macOS 应用', description: '连接系统日历与提醒事项，并使用原生权限模型。', category: '系统与效率', setupLabel: undefined },
+    'google-calendar': { name: 'Google 日历', description: '管理日程、创建会议并查询空闲时间。', category: '系统与效率', setupLabel: '需要 OAuth credentials 文件' },
+    figma: { name: 'Figma', description: '读取设计文件、组件与开发交付信息。', category: '设计与开发', setupLabel: '需要 Personal Access Token' },
+    vercel: { name: 'Vercel', description: '检查项目、部署状态、日志与平台文档。', category: '设计与开发', setupLabel: '需要登录授权' },
+    supabase: { name: 'Supabase', description: '管理数据库、项目配置、迁移与 Edge Functions。', category: '设计与开发', setupLabel: '需要登录授权' },
+    filesystem: { name: '本地文件', description: '在指定目录中安全地读取、写入和管理文件。', category: '文件与知识', setupLabel: '需要选择允许访问的目录' },
+    memory: { name: '持久记忆', description: '用结构化知识图谱记住实体、关系和重要事实。', category: '文件与知识', setupLabel: undefined },
+    playwright: { name: '浏览器自动化', description: '让 Maka 通过 Playwright 读取和操作真实网页。', category: '设计与开发', setupLabel: undefined },
+    'sequential-thinking': { name: '序列思考', description: '为复杂问题提供可修正、可验证的结构化推理。', category: '推理与规划', setupLabel: undefined },
+  },
+  'zh-TW': {
+    dingtalk: { name: 'DingTalk', description: '管理聯絡人、行事曆、任務和協作資料。', category: '通訊', setupLabel: '需要 Client ID 和 Client Secret' },
+    feishu: { name: '飛書', description: '存取飛書文件、行事曆、訊息和 OpenAPI。', category: '通訊', setupLabel: '需要 App ID 和 App Secret' },
+    slack: { name: 'Slack', description: '傳送訊息、管理頻道，並在 Slack 工作區協作。', category: '通訊', setupLabel: '需要 Bot Token 和 Team ID' },
+    line: { name: 'LINE', description: '透過 LINE Bot Messaging API 傳送和管理訊息。', category: '通訊', setupLabel: '需要 Channel Access Token' },
+    notion: { name: 'Notion', description: '搜尋、讀取和更新 Notion 工作區。', category: '知識與文件', setupLabel: '需要登入授權' },
+    'macos-apps': { name: 'macOS 應用程式', description: '透過原生系統權限連接「行事曆」與「提醒事項」。', category: '系統與生產力', setupLabel: undefined },
+    'google-calendar': { name: 'Google 日曆', description: '管理活動、建立會議並檢查可用時間。', category: '系統與生產力', setupLabel: '需要 OAuth 憑證檔案' },
+    figma: { name: 'Figma', description: '讀取設計檔案、元件和開發交付資料。', category: '設計與開發', setupLabel: '需要 Personal Access Token' },
+    vercel: { name: 'Vercel', description: '檢查專案、部署、記錄和平台文件。', category: '設計與開發', setupLabel: '需要登入授權' },
+    supabase: { name: 'Supabase', description: '管理資料庫、專案設定、移轉和 Edge Functions。', category: '設計與開發', setupLabel: '需要登入授權' },
+    filesystem: { name: '本機檔案', description: '在選取的資料夾中安全地讀取、寫入和管理檔案。', category: '檔案與知識', setupLabel: '需要選擇允許存取的資料夾' },
+    memory: { name: '持久記憶', description: '使用結構化知識圖譜記住實體、關係和重要事實。', category: '檔案與知識', setupLabel: undefined },
+    playwright: { name: '瀏覽器自動化', description: '讓 Maka 透過 Playwright 讀取和操作實際網頁。', category: '設計與開發', setupLabel: undefined },
+    'sequential-thinking': { name: '循序思考', description: '為複雜問題提供可修正、可驗證的結構化推理。', category: '推理與規劃', setupLabel: undefined },
+  },
+  en: {
+    dingtalk: { name: 'DingTalk', description: 'Manage contacts, calendars, tasks, and collaboration data.', category: 'Communication', setupLabel: 'Requires Client ID and Client Secret' },
+    feishu: { name: 'Feishu', description: 'Access Feishu documents, calendars, messages, and OpenAPI.', category: 'Communication', setupLabel: 'Requires App ID and App Secret' },
+    slack: { name: 'Slack', description: 'Send messages, manage channels, and collaborate in a Slack workspace.', category: 'Communication', setupLabel: 'Requires Bot Token and Team ID' },
+    line: { name: 'LINE', description: 'Send and manage messages through the LINE Bot Messaging API.', category: 'Communication', setupLabel: 'Requires Channel Access Token' },
+    notion: { name: 'Notion', description: 'Search, read, and update a Notion workspace.', category: 'Knowledge and documents', setupLabel: 'Requires sign-in authorization' },
+    'macos-apps': { name: 'macOS apps', description: 'Connect Calendar and Reminders through native system permissions.', category: 'System and productivity', setupLabel: undefined },
+    'google-calendar': { name: 'Google Calendar', description: 'Manage events, create meetings, and check availability.', category: 'System and productivity', setupLabel: 'Requires an OAuth credentials file' },
+    figma: { name: 'Figma', description: 'Read design files, components, and developer handoff data.', category: 'Design and development', setupLabel: 'Requires a Personal Access Token' },
+    vercel: { name: 'Vercel', description: 'Inspect projects, deployments, logs, and platform documentation.', category: 'Design and development', setupLabel: 'Requires sign-in authorization' },
+    supabase: { name: 'Supabase', description: 'Manage databases, project configuration, migrations, and Edge Functions.', category: 'Design and development', setupLabel: 'Requires sign-in authorization' },
+    filesystem: { name: 'Local files', description: 'Safely read, write, and manage files in selected directories.', category: 'Files and knowledge', setupLabel: 'Requires selecting allowed directories' },
+    memory: { name: 'Persistent memory', description: 'Remember entities, relationships, and important facts in a structured knowledge graph.', category: 'Files and knowledge', setupLabel: undefined },
+    playwright: { name: 'Browser automation', description: 'Let Maka read and operate real web pages through Playwright.', category: 'Design and development', setupLabel: undefined },
+    'sequential-thinking': { name: 'Sequential thinking', description: 'Provide revisable, verifiable structured reasoning for complex problems.', category: 'Reasoning and planning', setupLabel: undefined },
+  },
+} satisfies UiCatalog<Record<McpCatalogId, McpCatalogLocalizedCopy>>;
 
 export function getMcpCatalog(locale: UiLocale): McpCatalogEntry[] {
-  if (locale === 'zh-CN') return MCP_CATALOG;
-  const copy = locale === 'zh-TW' ? MCP_CATALOG_TRADITIONAL_COPY : MCP_CATALOG_ENGLISH_COPY;
+  const copy = MCP_CATALOG_COPY[locale];
   return MCP_CATALOG.map((entry) => ({ ...entry, ...copy[entry.id] }));
 }
