+++ apps/desktop/src/renderer/settings/runtime-host-profiles-section.tsx
@@ -49,6 +49,8 @@ import {
 } from "@maka/ui";
 import { Cpu, ICON_SIZE } from "@maka/ui/icons";
 import { getSettingsProjectsCopy } from "../locales/settings-projects-copy.js";
+import type { UiLocale } from "@maka/core/ui-locale";
+
 import { PasswordInput } from "./password-input.js";
 import { settingsActionErrorMessage } from "./settings-error-copy.js";
 import { SettingsField, SettingsRow, SettingsSection } from "./settings-section.js";
@@ -59,7 +61,6 @@ import {
 } from './runtime-host-management-dialog.js';
 import {
   PeerMeshPeerIdButton,
-  getRuntimeHostPeerMeshCopy,
   RuntimeHostAddComputerMenu,
   RuntimeHostConnectionCodeDialog,
   RuntimeHostPairingRecoveryButton,
@@ -93,7 +94,6 @@ export function RuntimeHostProfilesSection(props: {
 }) {
   const locale = useUiLocale();
   const copy = getSettingsProjectsCopy(locale).runtimeHost;
-  const peerMeshCopy = getRuntimeHostPeerMeshCopy(locale);
   const pairingActionCopy: RuntimeHostPairingActionCopy = {
     retry: copy.resolvePairingRecovery,
     retryFailed: copy.resolvePairingRecoveryFailed,
@@ -409,10 +409,10 @@ export function RuntimeHostProfilesSection(props: {
             localAccessEnabling
               ? copy.remoteAccessEnabling
               : localAccess?.state === 'unsupported'
-              ? localAccess.message
-              : localAccess?.state === 'unavailable'
                 ? localAccess.message
-                : copy.thisComputerRemoteAccessHelp
+                : localAccess?.state === 'unavailable'
+                  ? localAccess.message
+                  : copy.thisComputerRemoteAccessHelp
           }
           end={(
             <HStack gap={2} align="center">
@@ -670,9 +670,9 @@ export function RuntimeHostProfilesSection(props: {
                               <PeerMeshPeerIdButton
                                 peerId={profile.transport.reachability.lease.peerId}
                                 displayValue={abbreviatePeerId(profile.transport.reachability.lease.peerId)}
-                                copyLabel={peerMeshCopy.copyPeerId(profile.transport.reachability.lease.peerId)}
-                                copiedTitle={peerMeshCopy.peerIdCopied}
-                                failedTitle={peerMeshCopy.peerIdCopyFailed}
+                                copyLabel={copy.copyPeerId(profile.transport.reachability.lease.peerId)}
+                                copiedTitle={copy.peerIdCopied}
+                                failedTitle={copy.peerIdCopyFailed}
                                 errorMessage={(error) => settingsActionErrorMessage(error, locale)}
                                 className="settingsRuntimeHostPeerId"
                               />
@@ -695,13 +695,11 @@ export function RuntimeHostProfilesSection(props: {
                         <Badge variant="neutral" label={copy.unavailable} />
                       ) : null}
                       {entry.peerPath ? (
-                        <Tooltip content={peerPathDetail(entry.peerPath, peerMeshCopy)}>
+                        <Tooltip content={peerPathDetail(entry.peerPath, locale)}>
                           <span>
                             <Badge
                               variant="neutral"
-                              label={entry.peerPath.kind === 'direct'
-                                ? peerMeshCopy.peerPathDirect
-                                : peerMeshCopy.peerPathTransit}
+                              label={peerPathLabel(entry.peerPath, locale)}
                             />
                           </span>
                         </Tooltip>
@@ -844,8 +842,9 @@ export function RuntimeHostProfilesSection(props: {
 
 function peerPathDetail(
   path: RuntimeHostPeerConnectionPath,
-  copy: ReturnType<typeof getRuntimeHostPeerMeshCopy>,
+  locale: UiLocale,
 ): string {
+  const copy = getSettingsProjectsCopy(locale).runtimeHost;
   if (path.kind === 'transit') {
     return `${copy.peerPathTransit} · ${abbreviatePeerId(path.relayPeerId)}`;
   }
@@ -855,10 +854,23 @@ function peerPathDetail(
       ? 'QUIC'
       : path.transport === 'tcp'
         ? 'TCP'
-        : copy.peerPathOther;
+        : copy.peerPathTransportOther;
   return `${copy.peerPathDirect} · ${transport}`;
 }
 
+function peerPathLabel(
+  path: RuntimeHostPeerConnectionPath,
+  locale: UiLocale,
+): string {
+  const copy = getSettingsProjectsCopy(locale).runtimeHost;
+  if (path.kind === 'transit') {
+    return copy.peerPathTransit;
+  }
+  if (path.transport === 'webrtc') return 'WebRTC';
+  if (path.transport === 'quic') return 'QUIC';
+  if (path.transport === 'tcp') return 'TCP';
+  return copy.peerPathDirect;
+}
 
 function abbreviatePeerId(peerId: string): string {
   return peerId.length <= 20 ? peerId : `${peerId.slice(0, 10)}…${peerId.slice(-6)}`;
